Workaround for using snapshot with async rsc and vitest
I discovered what I believe is an issue when using snapshots with Vitest and React server components.
The snapshots were being generated as an empty `<div />` tag.
My workaround was to first perform some `screen` assertions (as shown in the test cases).
After doing this, the snapshots are generated correctly when using `toMatchSnapshot`.